Refactor so no longer contains voucher swap and improvements(sorry for the mad commit, a lot was going on)

+++ postexp/postexp.m
@@ -13,10 +13,9 @@
 #import <sys/stat.h>
 #import <sys/utsname.h>
 
-#include "kernel_memory.h"
+#include "kmem.h"
+#include "kernel_utils.h"
 #include "kernel_call.h"
-#include "parameters.h"
-#include "kernel_slide.h"
 #include "postexp.h"
 #include "offsets.h"
 #include "root.h"
@@ -26,12 +25,14 @@
 #include "launch_utils.h"
 #include "payload.h"
 #include "offsets_dump.h"
+#include "remap_tfp_set_hsp.h"
 
 #include "patchfinder64.h"
 #include "macho-helper.h"
 #include "lzssdec.hpp"
 #include "untar.h"
 #include "amfi_utils.h"
+#include "utils.h"
 
 #define in_bundle(obj) strdup([[[[NSBundle mainBundle] bundlePath] stringByAppendingPathComponent:@obj] UTF8String])
 
@@ -58,83 +59,61 @@ ERROR("error moving item %s to path %s (%s)", copyFrom, moveTo, [[error localize
 }
 
 const char *kernel_path = "/System/Library/Caches/com.apple.kernelcaches/kernelcache";
+bool static_kernel = false;
 
-enum post_exp_t clean_up_previous(void) {
-    NSError *error = NULL;
-    if (!fileExists("/var/containers/Bundle/.installed_rootlessJB3")) {
-        
-        if (fileExists("/var/containers/Bundle/iosbinpack64")) {
-            INFO("uninstalling previous build...");
-            
-            removeFile("/var/LIB");
-            removeFile("/var/ulb");
-            removeFile("/var/bin");
-            removeFile("/var/sbin");
-            removeFile("/var/containers/Bundle/tweaksupport/Applications");
-            removeFile("/var/Apps");
-            removeFile("/var/profile");
-            removeFile("/var/motd");
-            removeFile("/var/dropbear");
-            removeFile("/var/containers/Bundle/tweaksupport");
-            removeFile("/var/containers/Bundle/iosbinpack64");
-            removeFile("/var/containers/Bundle/dylibs");
-            removeFile("/var/log/testbin.log");
-            
-            if (fileExists("/var/log/jailbreakd-stdout.log")) removeFile("/var/log/jailbreakd-stdout.log");
-            if (fileExists("/var/log/jailbreakd-stderr.log")) removeFile("/var/log/jailbreakd-stderr.log");
+enum post_exp_t recover_with_hsp4(bool use_static_kernel, uint64_t *ext_kernel_slide, uint64_t *ext_kernel_load_base) {
+    struct task_dyld_info dyld_info = { 0 };
+    mach_msg_type_number_t count = TASK_DYLD_INFO_COUNT;
+    if((host_get_special_port(mach_host_self(), HOST_LOCAL_NODE, 4, &kernel_task_port) == KERN_SUCCESS) && MACH_PORT_VALID(kernel_task_port)) {
+        if(task_info(kernel_task_port, TASK_DYLD_INFO, (task_info_t)&dyld_info, &count) == KERN_SUCCESS) {
+            kernel_load_base = dyld_info.all_image_info_addr;
+            kernel_slide = dyld_info.all_image_info_size;
+            *ext_kernel_slide = kernel_slide;
+            *ext_kernel_load_base = kernel_load_base;
+            return NO_ERROR;
         }
-        
-        INFO("installing bootstrap...");
-        
-        chdir("/var/containers/Bundle/");
-        FILE *bootstrap = fopen((char*)in_bundle("iosbinpack.tar"), "r");
-        untar(bootstrap, "/var/containers/Bundle/");
-        fclose(bootstrap);
-        
-//        FILE *tweaks = fopen((char*)in_bundle("tweaksupport.tar"), "r");
-//        untar(tweaks, "/var/containers/Bundle/");
-//        fclose(tweaks);
-        
-//        if(!fileExists("/var/containers/Bundle/tweaksupport") || !fileExists("/var/containers/Bundle/iosbinpack64")) {
-//            ERROR("[-] Failed to install bootstrap");
-//        }
-        
-        mkdir("/var/containers/Bundle/tweaksupport", 0777);
-        if(!fileExists("/var/containers/Bundle/iosbinpack64")) {
-            ERROR("failed to install bootstrap");
-            cleanup();
-            return ERROR_INSTALLING_BOOTSTRAP;
-        }
-        
-        INFO("creating symlinks...");
-        
-//        symlink("/var/containers/Bundle/tweaksupport/Library", "/var/LIB");
-//        symlink("/var/containers/Bundle/tweaksupport/usr/lib", "/var/ulb");
-//        symlink("/var/containers/Bundle/tweaksupport/Applications", "/var/Apps");
-//        symlink("/var/containers/Bundle/tweaksupport/bin", "/var/bin");
-//        symlink("/var/containers/Bundle/tweaksupport/sbin", "/var/sbin");
-//        symlink("/var/containers/Bundle/tweaksupport/usr/libexec", "/var/libexec");
-        
-        close(open("/var/containers/Bundle/.installed_rootlessJB3", O_CREAT));
-        
-        //limneos
-        symlink("/var/containers/Bundle/iosbinpack64/etc", "/var/etc");
-//        symlink("/var/containers/Bundle/tweaksupport/usr", "/var/usr");
-        symlink("/var/containers/Bundle/iosbinpack64/usr/bin/killall", "/var/bin/killall");
-        
-        INFO("installed bootstrap!");
     }
-    return NO_ERROR;
+    return ERROR_TFP0_NOT_RECOVERED;
 }
 
-enum post_exp_t root_and_escape(void) {
+enum post_exp_t init(mach_port_t tfp0, bool use_static_kernel, uint64_t *ext_kernel_slide, uint64_t *ext_kernel_load_base) {
     // Initialize offsets
     _offsets_init();
     
+    kernel_task_port = tfp0;
+    static_kernel = use_static_kernel;
+    if((*ext_kernel_slide != 0) && (*ext_kernel_load_base != 0)) {
+        kernel_load_base = *ext_kernel_load_base;
+        kernel_slide = *ext_kernel_slide;
+    } else if((*ext_kernel_slide == 0) && (*ext_kernel_load_base != 0)) {
+        kernel_load_base = *ext_kernel_load_base;
+        kernel_slide = kernel_load_base - kernel_base;
+        *ext_kernel_slide = kernel_slide;
+    } else if((*ext_kernel_slide != 0) && (*ext_kernel_load_base == 0)) {
+        kernel_slide = *ext_kernel_slide;
+        kernel_load_base = kernel_base + kernel_slide;
+        *ext_kernel_load_base = kernel_load_base;
+    } else {
+        kernel_load_base = find_kernel_base();
+        kernel_slide = kernel_load_base - kernel_base;
+        *ext_kernel_slide = kernel_slide;
+        *ext_kernel_load_base = kernel_load_base;
+    }
+    
+    if (!verify_tfp0()) {
+        ERROR("Failed to verify TFP0.");
+        return ERROR_INITIALAZING_LIBRARY;
+    }
+
+    return NO_ERROR;
+}
+
+enum post_exp_t root_pid(pid_t pid) {
+    uint64_t task_struct = task_struct_of_pid(pid);
     // Get r00t
-    save_proc_user_struct(current_task);
+    save_proc_user_struct(task_struct);
     INFO("current UID: %d", getuid());
-    root(current_task);
+    root(task_struct);
     uid_t current_uid = getuid();
     if(current_uid != 0) {
         ERROR("couldn't get r00t");
@@ -143,13 +122,17 @@ enum post_exp_t root_and_escape(void) {
     } else {
         INFO("current UID: %d", getuid());
     }
-    
+    return NO_ERROR;
+}
+
+enum post_exp_t unsandbox_pid(pid_t pid) {
+    uint64_t task_struct = task_struct_of_pid(pid);
     // Unsandbox
-    save_proc_sandbox_struct(current_task);
-    unsandbox(current_task);
+    save_proc_sandbox_struct(task_struct);
+    unsandbox(task_struct);
     
-    setcsflags(current_task);
-    platformize(current_task);
+    setcsflags(task_struct);
+    platformize(task_struct);
     INFO("the application is now a platform binary");
     
     return NO_ERROR;
@@ -161,7 +144,7 @@ enum post_exp_t get_kernel_file(void) {
     
     const char *location = [[docs stringByAppendingPathComponent:[NSString stringWithFormat:@"kernelcache.dump"]] UTF8String];
     NSError *error = NULL;
-        
+    
     removeFile(location);
     error = NULL;
     copyFile(kernel_path, location);
@@ -174,35 +157,62 @@ enum post_exp_t get_kernel_file(void) {
 }
 
 enum post_exp_t initialize_patchfinder64() {
-    NSString *docs = [[[[NSFileManager defaultManager] URLsForDirectory:NSDocumentDirectory inDomains:NSUserDomainMask] lastObject] path];
-    const char *original_kernel_cache_path = [[docs stringByAppendingPathComponent:[NSString stringWithFormat:@"kernelcache.dump"]] UTF8String];
-    const char *decompressed_kernel_cache_path = [[docs stringByAppendingPathComponent:[NSString stringWithFormat:@"kernelcache.dec"]] UTF8String];
-    
-    NSError *error = NULL;
-    removeFile(decompressed_kernel_cache_path);
-    
-    FILE *original_kernel_cache = fopen(original_kernel_cache_path, "rb");
-    uint32_t macho_header_offset = find_macho_header(original_kernel_cache);
-    char *args[5] = { "lzssdec", "-o", (char *)[NSString stringWithFormat:@"0x%x", macho_header_offset].UTF8String, (char *)original_kernel_cache_path, (char *)decompressed_kernel_cache_path};
-    lzssdec(5, args);
-    fclose(original_kernel_cache);
-    chown(decompressed_kernel_cache_path, 501, 501);
-    
-    if (init_kernel(NULL, 0, decompressed_kernel_cache_path) != ERR_SUCCESS) {
-        ERROR("failed to initialize patchfinder");
-        cleanup();
-        return ERROR_SETTING_PATCHFINDER64;
+    if(static_kernel) {
+        NSString *docs = [[[[NSFileManager defaultManager] URLsForDirectory:NSDocumentDirectory inDomains:NSUserDomainMask] lastObject] path];
+        const char *original_kernel_cache_path = [[docs stringByAppendingPathComponent:[NSString stringWithFormat:@"kernelcache.dump"]] UTF8String];
+        const char *decompressed_kernel_cache_path = [[docs stringByAppendingPathComponent:[NSString stringWithFormat:@"kernelcache.dec"]] UTF8String];
+        
+        NSError *error = NULL;
+        removeFile(decompressed_kernel_cache_path);
+        
+        FILE *original_kernel_cache = fopen(original_kernel_cache_path, "rb");
+        uint32_t macho_header_offset = find_macho_header(original_kernel_cache);
+        char *args[5] = { "lzssdec", "-o", (char *)[NSString stringWithFormat:@"0x%x", macho_header_offset].UTF8String, (char *)original_kernel_cache_path, (char *)decompressed_kernel_cache_path};
+        lzssdec(5, args);
+        fclose(original_kernel_cache);
+        chown(decompressed_kernel_cache_path, 501, 501);
+        
+        if (init_kernel(NULL, 0, decompressed_kernel_cache_path) != ERR_SUCCESS) {
+            ERROR("failed to initialize patchfinder");
+            cleanup();
+            return ERROR_SETTING_PATCHFINDER64;
+        } else {
+            INFO("patchfinder initialized successfully");
+            set_cached_offsets(kCFCoreFoundationVersionNumber);
+            term_kernel();
+            INFO("offsets dumped correctly and patchfinder terminated");
+            return NO_ERROR;
+        }
     } else {
-        INFO("patchfinder initialized successfully");
-        return NO_ERROR;
+        if (init_kernel(kread, kernel_load_base, NULL) != ERR_SUCCESS) {
+            ERROR("failed to initialize patchfinder");
+            cleanup();
+            return ERROR_SETTING_PATCHFINDER64;
+        } else {
+            INFO("patchfinder initialized successfully");
+            set_cached_offsets(kCFCoreFoundationVersionNumber);
+            term_kernel();
+            INFO("offsets dumped correctly and patchfinder terminated");
+            return NO_ERROR;
+        }
     }
 }
 
-enum post_exp_t bootstrap() {
-    enum post_exp_t clean_result = clean_up_previous();
-    if(clean_result != NO_ERROR) {
+enum post_exp_t set_host_special_port_4_patch(void) {
+    //---- host special port 4 ----//
+    if(setHSP4()) {
+        ERROR("failed to set tfp0 as hsp4!");
         cleanup();
-        return clean_result;
+        return ERROR_SETTING_HSP4;
+    }
+    return NO_ERROR;
+}
+
+enum post_exp_t bootstrap() {
+    current_task = task_struct_of_pid(getpid());
+    if(!clean_up_previous()) {
+        cleanup();
+        return ERROR_INSTALLING_BOOTSTRAP;
     }
     
     if(dump_offsets_to_file("/var/containers/Bundle/tweaksupport/offsets.data") != 0) {
@@ -256,7 +266,6 @@ enum post_exp_t bootstrap() {
     chmod("/var/motd", 0777);
     
     launch("/var/containers/Bundle/iosbinpack64/usr/bin/killall", "-SEGV", "dropbear", NULL, NULL, NULL, NULL, NULL);
-//    launch_as_platform("/var/containers/Bundle/iosbinpack64/usr/local/bin/dropbear", "-R", "-E", "-p", "22", "-p", "2222", NULL);
     
     if(fileExists(in_bundle("dropbear.plist"))) {
         removeFile("/var/containers/Bundle/iosbinpack64/LaunchDaemons/dropbear.plist");
@@ -313,7 +322,9 @@ enum post_exp_t bootstrap() {
 
 void cleanup(void) {
     INFO("cleaning up");
-    term_kernel();
+    if (verify_tfp0() && cached_offsets.allproc && !current_task) {
+        current_task = task_struct_of_pid(getpid());
+    }
     restore_csflags(current_task);
     sandbox(current_task);
     unroot(current_task);
